// features/chakra/types.ts
export type Chakra = {
  slug: string;
  name: string;
  sanskrit: string;
  meaning: string;
  location: string;
  color: string;
  element: string;
  sense: string;
  seedMantra: string;
  petals: number;
  description: string;
  qualitiesBalanced: string[];
  qualitiesImbalanced: string[];
  affirmation: string;
  healingPractices: string[];
};